Move sidebar for chart into sidebar dir
Eventually, we'll be moving more chart-specific controls into this
sidebar (toggling lines on/off).
